refactor: split core rendering and font logic into separate modules
- Removed `main.c` and created new entrypoint `pretty.c`
- Added `font.c` / `font.h` to encapsulate font discovery & loading
  - Implemented `find_font_path_from_fc_name()` using Fontconfig
  - Added `collect_font()` for initializing and validating monospace fonts
- Added `renderer.c` / `renderer.h` to handle frame rendering & FPS display
- Moved HEX color conversion macros into `macro_utils.h`
- Introduced `config_paths` struct in `config.h`
- Updated `pretty.h` to expose `notify_ui_flush()`
- Overall cleaned up `main.c` by modularizing rendering and font-related logic
